+++ examples/server/src/phecda.d.ts
@@ -1,0 +1,10 @@
+import 'phecda-server'
+
+declare module 'phecda-server'{
+
+     
+      interface PhecdaEvents{
+      watch:number
+      }
+}
+