+++ packages/core/src/custom/decorator.ts
@@ -2,8 +2,9 @@
 
 import { regisHandler, setModalVar } from '../core'
 import { getProperty } from '../namespace'
+import type { PhecdaEvents } from '../types'
 
-export function Watcher(eventName: string, options?: { once: boolean }) {
+export function Watcher(eventName: keyof PhecdaEvents, options?: { once: boolean }) {
   return (obj: any, key: string) => {
     setModalVar(obj, key)
     regisHandler(obj, key, {